Split prior module into prior evaulator module and distributions module.
