+++ pycbc/inference/__init__.py
@@ -1,3 +1,4 @@
+from pycbc.inference.distributions import *
 from pycbc.inference.likelihood import *
 from pycbc.inference.sampler import *
 from pycbc.inference.prior import *
